updated author csv and xq
there’s a namespace issue right now in the code. had to drop the oai:
namespace from the record element to get the author csv to work. going
to go and do this for the other files as well. I think the issue has to
do with how the data is passed into the DB via the get-oak-data script.